Bug fix
When a zoneplayer leaves a group, but has no input selected itself, the _Sonos_ accessory would often show _On_ (play/pause) as true.  This is because the GroupRenderingControl event for the Coordinator would arrive before the GroupManagement event for the former member, so the coordinator settings are copied back to the (now former) member.  When the former member has input selected, it receives an AVTransport event that overwrites the copied settings.